+++ apps/dashboard/src/app/entities/entity-dialog/group-select/group-select.component.ts
@@ -1,7 +1,7 @@
-import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { stfAnimations } from '@flaps/pastanaga';
 import { Entities } from '@nuclia/core';
-import { GROUP_COLORS } from '../../model';
+import { generatedEntitiesColor } from '../../model';
 
 @Component({
   selector: 'app-group-select',
@@ -17,7 +17,7 @@ export class GroupSelectComponent {
   @Input() disabled: boolean = false;
 
   open: boolean = false;
-  colors = GROUP_COLORS;
+  colors = generatedEntitiesColor;
 
   constructor() {}