+++ apps/dashboard/src/app/resources/edit/edit-resource.helpers.spec.ts
@@ -1,17 +1,18 @@
 import { FIELD_TYPE, FieldId } from '@nuclia/core';
-import { getUpdatedUserFieldMetadata } from './edit-resource.helpers';
+import { getFieldMetadataForClassifications } from './edit-resource.helpers';
 
 describe('Edit resource helpers', () => {
   const field: FieldId = { field_id: 'f1', field_type: FIELD_TYPE.text };
 
   it('should add a new label when none exists', () => {
-    const all = getUpdatedUserFieldMetadata(
+    const all = getFieldMetadataForClassifications(
       field,
       [
         {
           paragraphId: 'p1',
           userClassifications: [{ labelset: 'heroes', label: 'batman' }],
           generatedClassifications: [],
+          text: '',
         },
       ],
       [],
@@ -25,7 +26,7 @@ describe('Edit resource helpers', () => {
   });
 
   it('should update labels and keep tokens as well as other fields', () => {
-    const all = getUpdatedUserFieldMetadata(
+    const all = getFieldMetadataForClassifications(
       field,
       [
         {
@@ -35,16 +36,19 @@ describe('Edit resource helpers', () => {
             { labelset: 'heroes', label: 'catwoman' },
           ],
           generatedClassifications: [],
+          text: '',
         },
         {
           paragraphId: 'p2',
           userClassifications: [{ labelset: 'heroes', label: 'catwoman' }],
           generatedClassifications: [],
+          text: '',
         },
         {
           paragraphId: 'p3',
           userClassifications: [],
           generatedClassifications: [],
+          text: '',
         },
       ],
       [